working multiplication implementation
